dev/requirements.txt
# Requirements for all the Python build scripts

# Protoc is required for the codegen scripts which work off the API proto files
protoc-wheel-0 >= 3.13.0, < 3.14.0
protobuf >= 3.13.0, < 3.14.0
